requirements.txt
streamlit
langchain
langchain-openai
python-dotenv


